+++ packages/sdk/src/urls.ts
@@ -7,12 +7,14 @@ export const URLS = {
   },
   rest: {
     documents: {
-      pick: (opts: Pick<MCDocument, 'documentID'> | Pick<MCDocument, 'documentID' | 'major' | 'minor'>) => {
-        const {documentID} = opts;
-        let queryParams = "";
+      pick: (
+        opts: Pick<MCDocument, 'documentID'> | Pick<MCDocument, 'documentID' | 'major' | 'minor'>,
+      ) => {
+        const { documentID } = opts;
+        let queryParams = '';
 
         if ('major' in opts) {
-          const {major, minor} = opts;
+          const { major, minor } = opts;
 
           queryParams = `v${major ?? 0}.${minor ?? 1}`;
         }
@@ -21,9 +23,9 @@ export const URLS = {
         return {
           presentations: `${baseURL}/presentations`,
           self: baseURL,
-          withVersion: `${baseURL}${queryParams}`
+          withVersion: `${baseURL}${queryParams}`,
         };
-      }
+      },
     },
     users: {
       self: `/rest-api/users/me`,