+++ packages/sdk/CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changes
+
+- `MermaidChart#resetAccessToken()` no longer returns a `Promise`.
+
 ## [0.2.0] - 2024-04-11
 
 ### Added